Moved shell scripts into separate directory
