+++ pipeline/pipeline_stack.py
@@ -1,15 +1,13 @@
 from constructs import Construct
 from aws_cdk import (
     Stack,
-    aws_codepipeline as codepipeline,
-    aws_codepipeline_actions as cpactions,
     pipelines as pipelines,
     aws_codedeploy as codedeploy,
 )
-import jsii
 
 from .backend_stage import ServerlessTodoBackendStage
 from .lib.codebuild_execution_role import CodeBuildExecutionRole
+from .lib.bluegreen_deployment_step import BlueGreenDeploymentStep
 from .web_stage import ServerlessTodoWebStage
 
 
@@ -17,11 +15,14 @@ class ServerlessTodoPipelineStack(Stack):
     def __init__(self, scope: Construct, id: str, **kwargs) -> None:
         super().__init__(scope, id, **kwargs)
 
+        prefix = "Todo"
+
         ############################ PIPELINE START ############################
         pipeline = pipelines.CodePipeline(
             self,
-            "TodoPipeline",
-            pipeline_name="TodoPipeline",
+            f"{prefix}Pipeline",
+            pipeline_name=f"{prefix}Pipeline",
+            publish_assets_in_parallel=False,
             synth=pipelines.ShellStep(
                 "Synth",
                 input=pipelines.CodePipelineSource.git_hub(
@@ -29,39 +30,30 @@ class ServerlessTodoPipelineStack(Stack):
                     "feature/fargate-web-app",
                 ),
                 commands=[
-                    "npm install -g aws-cdk",
-                    "pip install -r requirements.txt",
-                    # TODO: "pytest tests"
-                    "cdk synth",
-                    # We need to include the codedeploy stuff in our pipeline's cloud assembly file set
-                    "cp -r ./scripts/codedeploy cdk.out",
+                    "chmod a+x ./scripts/pipeline/synth",
+                    "./scripts/pipeline/synth",
                 ],
             ),
-            publish_assets_in_parallel=False,
         )
         ############################ PIPELINE END ############################
 
         ############################ APPLICATION / STORAGE START ############################
         backend = ServerlessTodoBackendStage(
-            self, "TodoBackendStage", prefix="Todo", **kwargs
+            self, f"{prefix}BackendStage", prefix=prefix, **kwargs
         )
         pipeline.add_stage(backend)
         ############################ APPLICATION / STORAGE END ############################
 
         ############################ WEB START ############################
-        container_repository = (
-            "460848972690.dkr.ecr.eu-west-2.amazonaws.com/serverless-todo-web-app"
-        )
-
         configure_codedeploy_step = pipelines.ShellStep(
             "ConfigureBlueGreenDeployment",
             input=pipeline.cloud_assembly_file_set,
             primary_output_directory="codedeploy",
             commands=[
                 "ls",
-                "chmod a+x ./codedeploy/codedeploy_configuration.sh",
+                "chmod a+x ./codedeploy/configure_codedeploy_step",
                 (
-                    "./codedeploy/codedeploy_configuration.sh "
+                    "./codedeploy/configure_codedeploy_step "
                     f"{self.account} "
                     f"{self.region} "
                     f"{'Todo'} "
@@ -87,11 +79,15 @@ class ServerlessTodoPipelineStack(Stack):
             deployment_config=codedeploy.EcsDeploymentConfig.CANARY_10_PERCENT_5_MINUTES,
         )
 
-        deploy_step = CodeDeployStep(
+        deploy_step = BlueGreenDeploymentStep(
             input=configure_codedeploy_step.primary_output,
             deployment_group=cd_deployment_group,
         )
         deploy_step.add_step_dependency(configure_codedeploy_step)
+
+        container_repository = (
+            "460848972690.dkr.ecr.eu-west-2.amazonaws.com/serverless-todo-web-app"
+        )
 
         pipeline.add_stage(
             ServerlessTodoWebStage(
@@ -102,61 +98,21 @@ class ServerlessTodoPipelineStack(Stack):
                 **kwargs,
             ),
             pre=[
-                # TODO: change this to a generic ShellStep and add the role to the pipeline itself?
+                # TODO change this to a generic ShellStep and add the role to the pipeline itself?
                 pipelines.CodeBuildStep(
                     "BuildAndUploadDockerImage",
                     role=CodeBuildExecutionRole(
                         self, "TodoCodeBuildExecutionRole"
                     ).role,
+                    # This step uses the pipeline's source file set by default (i.e. GitHub),
+                    # meaning we don't have to copy the web directory in the pipeline's synth step
+                    # like we do the scripts directory.
                     commands=[
-                        "cd web/",
-                        "echo Logging in to Amazon ECR...",
-                        f"aws ecr get-login-password | docker login --username AWS --password-stdin {container_repository}",
-                        "docker build -t todo-web-app .",
-                        f"docker tag todo-web-app:latest {container_repository}:latest",
-                        f"docker push {container_repository}:latest",
+                        "chmod a+x ./scripts/pipeline/push_to_ecr",
+                        f"./scripts/pipeline/push_to_ecr {container_repository}",
                     ],
                 )
             ],
             post=[configure_codedeploy_step, deploy_step],
         )
         ############################ WEB END ############################
-
-
-@jsii.implements(pipelines.ICodePipelineActionFactory)
-class CodeDeployStep(pipelines.Step):
-    def __init__(self, input: pipelines.FileSet, deployment_group):
-        super().__init__("MyCodeDeployStep")
-
-        self._input = input
-        self._deployment_group = deployment_group
-
-    def produce_action(
-        self,
-        stage: codepipeline.IStage,
-        *,
-        scope,
-        action_name,
-        run_order,
-        artifacts: pipelines.ArtifactMap,
-        pipeline,
-    ):
-        artifact = artifacts.to_code_pipeline(self._input)
-
-        action = cpactions.CodeDeployEcsDeployAction(
-            action_name="CustomCodeDeployAction",
-            deployment_group=self._deployment_group,
-            app_spec_template_input=artifact,
-            task_definition_template_input=artifact,
-            run_order=run_order,
-            container_image_inputs=[
-                cpactions.CodeDeployEcsContainerImageInput(
-                    input=artifact,
-                    task_definition_placeholder="IMAGE1_NAME",
-                )
-            ],
-        )
-
-        stage.add_action(action)
-
-        return pipelines.CodePipelineActionFactoryResult(run_orders_consumed=1)
