+++ scripts/README.md
@@ -1,0 +1,5 @@
+## Scripts
+
+Within this directory is a collection of utlity scripts written either for developer covenience, or for use by the CICD pipeline.
+
+All scripts are written with the intention of running with the project root as the current working directory, hence all files and paths should be considered relative to the project root.
