+++ scripts/pipeline/push_to_ecr
@@ -1,0 +1,20 @@
+#!/bin/sh
+#
+# This script containerises the web app using Docker, tags it as "latest" and
+# pushes it to ECR for our CodeDeploy job to later source during the blue /
+# green deployment.
+#
+# This will happend even if there are no changes in the web app's source code.
+#
+
+repo=$1
+echo "ECR Repository: "$repo
+
+cd web/ || exit
+
+echo Logging in to Amazon ECR,
+aws ecr get-login-password | docker login --username AWS --password-stdin "$repo"
+
+docker build -t web-app .
+docker tag web-app:latest "$repo":latest
+docker push "$repo":latest
